Moved BlendMode, Color, ColorDepth, LoopDirection, Point, Rectangle, and Size into the AsepriteDotNet.Common namespace
