+++ source/AsepriteDotNet/AsepriteFile.cs
@@ -19,7 +19,7 @@ IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN
 CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 ----------------------------------------------------------------------------- */
 using System.Collections.ObjectModel;
-
+using AsepriteDotNet.Common;
 using AsepriteDotNet.Document;
 using AsepriteDotNet.Image;
 using AsepriteDotNet.IO;
