+++ source/AsepriteDotNet/Common/LoopDirection.cs
@@ -18,7 +18,7 @@ COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER
 IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN
 CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 ----------------------------------------------------------------------------- */
-namespace AsepriteDotNet;
+namespace AsepriteDotNet.Common;
 
 /// <summary>
 ///     Defines the values that describe an animation loop direction.
