+++ source/AsepriteDotNet/Common/Size.cs
@@ -20,7 +20,7 @@ CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 ----------------------------------------------------------------------------- */
 using System.Diagnostics.CodeAnalysis;
 
-namespace AsepriteDotNet;
+namespace AsepriteDotNet.Common;
 
 /// <summary>
 ///     Represents the width and height of something.
