+++ source/AsepriteDotNet/Document/GroupLayer.cs
@@ -24,6 +24,7 @@ SOFTWARE.
 using System.Collections;
 using System.Collections.ObjectModel;
 using System.Diagnostics;
+using AsepriteDotNet.Common;
 
 namespace AsepriteDotNet.Document;
