+++ source/AsepriteDotNet/Document/ImageCel.cs
@@ -21,6 +21,7 @@ LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 SOFTWARE.
 ---------------------------------------------------------------------------- */
+using AsepriteDotNet.Common;
 using AsepriteDotNet.IO.Image;
 
 namespace AsepriteDotNet.Document;
