+++ source/AsepriteDotNet/Document/UserData.cs
@@ -19,7 +19,7 @@ IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN
 CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 ----------------------------------------------------------------------------- */
 using System.Diagnostics.CodeAnalysis;
-
+using AsepriteDotNet.Common;
 namespace AsepriteDotNet.Document;
 
 /// <summary>
