+++ source/AsepriteDotNet/IO/AsepriteFileReader.cs
@@ -22,7 +22,7 @@ OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 SOFTWARE.
 ---------------------------------------------------------------------------- */
 using System.Diagnostics;
-
+using AsepriteDotNet.Common;
 using AsepriteDotNet.Compression;
 using AsepriteDotNet.Document;
