+++ source/AsepriteDotNet/Image/Spritesheet.cs
@@ -22,6 +22,7 @@ OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 SOFTWARE.
 ---------------------------------------------------------------------------- */
 using System.Collections.ObjectModel;
+using AsepriteDotNet.Common;
 using AsepriteDotNet.IO.Image;
 
 namespace AsepriteDotNet.Image;
