+++ source/AsepriteDotNet/Image/SpritesheetAnimation.cs
@@ -22,6 +22,7 @@ OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 SOFTWARE.
 ---------------------------------------------------------------------------- */
 using System.Collections.ObjectModel;
+using AsepriteDotNet.Common;
 
 namespace AsepriteDotNet.Image;
