+++ source/AsepriteDotNet/Image/Tilesheet.cs
@@ -23,6 +23,7 @@ SOFTWARE.
 ---------------------------------------------------------------------------- */
 using System.Collections;
 using System.Collections.ObjectModel;
+using AsepriteDotNet.Common;
 using AsepriteDotNet.IO.Image;
 
 namespace AsepriteDotNet.Image;
